// client/src/Interfaces/RelationsState.tsx
export interface RelationsState {
  PropertyCategory: boolean;
  PropertyCondition: boolean;
  PropertyCountry: boolean;
  PropertyGarden: boolean;
  PropertyImage: boolean;
  PropertyPrice: boolean;
  PropertyServices: boolean;
  PropertyState: boolean;
  PropertyZone: boolean;
  PropertyUser: boolean;
  Price: boolean;
  Zone: boolean;
  loading: boolean;
  error: string | null;
}